only export issueFirstLastReleaseQueryOptions
